full_pr11

+++ src/utils/api.js
@@ -41,24 +41,23 @@ export default class Api {
       })
   }
 
-  deleteCardLike(cardId) {
-    return fetch(`${this._options.baseUrl}/cards/${cardId}/likes`, {
-      method: 'DELETE',
-      headers: this._options.headers
-    })
-      .then(res => {
-        return this._getServerStatus(res)
+  changeLikeCardStatus(cardId, isLiked) {
+    if (isLiked)
+      return fetch(`${this._options.baseUrl}/cards/${cardId}/likes`, {
+        method: 'DELETE',
+        headers: this._options.headers
       })
-  }
-
-  putCardLike(cardId) {
-    return fetch(`${this._options.baseUrl}/cards/${cardId}/likes`, {
-      method: 'PUT',
-      headers: this._options.headers
-    })
-      .then(res => {
-        return this._getServerStatus(res)
+        .then(res => {
+          return this._getServerStatus(res)
+        })
+    else 
+      return fetch(`${this._options.baseUrl}/cards/${cardId}/likes`, {
+        method: 'PUT',
+        headers: this._options.headers
       })
+        .then(res => {
+          return this._getServerStatus(res)
+        })
   }
 
   getUserInfo() {
